// frontend/src/app/shared/generated-api/model/exportRequest.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExportRequest { 
    scope: ExportRequestScopeEnum;
    resourceId?: string;
}
export enum ExportRequestScopeEnum {
    All = 'ALL',
    Notebook = 'NOTEBOOK',
    Note = 'NOTE'
};